Refactor package
